\name{LVQTools-package}
\alias{LVQTools-package}
\alias{LVQTools}
\docType{package}
\title{
Learning Vector Quantization Tools
}
\description{
This package provides several LVQ algorithms along with different usage methods and output visualization.
}
\details{
\tabular{ll}{
Package: \tab LVQTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-14\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
  See package documentation: LVQTools Bachelor project: implementing LVQ in R, LVQTools Usage: implementing LVQ in R,
			      LVQTools Output Usage: implementing LVQ in R, LVQTools Documentation: implementing LVQ in R.
}
\author{
Sander Kelders

Maintainer: Who to complain to <sakelders@gmail.com>
}
\references{
LVQTools Bachelor project: implementing LVQ in R
}

\seealso{
	  \code{\link{validate}}
	  \code{\link{show}},
	  \code{\link{showAll}},
	  \code{\link{getCostcurve}},
	  \code{\link{getPrototypeProgress}},
	  \code{\link{getPrototypes}},
	  \code{\link{getRelevanceProgress}},
	  \code{\link{getRelevances}},
	  \code{\link{getTestErrorProgress}},
	  \code{\link{getTestError}},
	  \code{\link{getTrainErrorProgress}},
	  \code{\link{getTrainError}},
	  \code{\link{nfoldoutput-class}},
	  \code{\link{trainoutput-class}},
	  \code{\link{traintestoutput-class}}
}

