\name{showAll}
\alias{showAll}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ showAll }
\description{
  This function shows all the available ouptut, by calling show with all the appropriate parameters to TRUE.
}
\usage{
showAll(LVQoutput)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{LVQoutput}{ The output-class containing all the output. }
}

\value{
  Graphics and/or values printed to screen depending on the output object.
}
\references{                 LVQTools Bachelor project: implementing LVQ in R
 }
\author{ Sander Kelders }

\seealso{
	  \code{\link{validate}}
	  \code{\link{showStuff}},
	  \code{\link{showAll}},
	  \code{\link{getCostcurve}},
	  \code{\link{getPrototypeProgress}},
	  \code{\link{getPrototypes}},
	  \code{\link{getRelevanceProgress}},
	  \code{\link{getRelevances}},
	  \code{\link{getTestErrorProgress}},
	  \code{\link{getTestError}},
	  \code{\link{getTrainErrorProgress}},
	  \code{\link{getTrainError}},
	  \code{\link{nfoldoutput-class}},
	  \code{\link{trainoutput-class}},
	  \code{\link{traintestoutput-class}}
}


